完成函数框架逻辑；修改数据类型。
空间不足，数据类型改为float；并增加宏FORMAT用于C输入输出。

+++ large-scale-classification/SVM.h
@@ -14,7 +14,7 @@ class SVM {
 
  public:
 
-    typedef double data_t;
+    typedef float data_t;
 
     static void train(const data_t * const *yx, const int l, const int n, data_t *bw);
     static void predict(data_t **yx, const int l, const int n, data_t *bw);
